Add postgres and nginx images

+++ recipes/happiness-service.rb
@@ -11,15 +11,17 @@ include_recipe 'docker'
 
 docker_image 'austenito/ruby-2.1.2'
 
-execute('stop container') { command "docker stop -t 60 happiness-service" }
-execute('remove container') { command "docker rm -f happiness-service" }
+if `sudo docker ps -a | grep happiness-service`.size > 0
+  execute('stop container') { command "docker stop -t 60 happiness-service" }
+  execute('remove container') { command "docker rm -f happiness-service" }
+end
 
 docker_container 'happiness-service' do
   image 'austenito/ruby-2.1.2'
   container_name "happiness-service"
   detach true
   env ["LOGENTRIES_HAPPINESS_SERVICE_TOKEN=#{node['logentries']['happiness_service']}"]
-  link ['postgres-production:db']
+  link ['postgres:db']
   volumes_from 'happiness-data'
   action :run
   port '3000:3000'
