+++ recipes/happiness.rb
@@ -10,8 +10,10 @@ include_recipe 'docker'
 
 docker_image 'austenito/ruby-2.1.2'
 
-execute('stop container') { command "docker stop -t 60 happiness" }
-execute('remove container') { command "docker rm -f happiness" }
+if `sudo docker ps -a | grep happiness,`.size > 0
+  execute('stop container') { command "docker stop -t 60 happiness" }
+  execute('remove container') { command "docker rm -f happiness" }
+end
 
 docker_container 'happiness' do
   image 'austenito/ruby-2.1.2'
@@ -20,7 +22,7 @@ docker_container 'happiness' do
   env ["LOGENTRIES_HAPPINESS_TOKEN=#{node['logentries']['happiness_service']}",
        "POPTART_API_TOKEN=#{node['poptart']['token']}"
       ]
-  link ['postgres-production:db', 'happiness-service:happiness_service']
+  link ['postgres:db', 'happiness-service:happiness_service']
   volumes_from 'happiness-data'
   action :run
   port '3001:3001'
